+++ src/main/java/racing/view/ResultView.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.view;
 
 import java.util.List;
