+++ src/test/java/racing/CarTest.java
@@ -4,6 +4,7 @@ package racing;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
