+++ src/test/java/racing/DiceTest.java
@@ -2,6 +2,7 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.util.Dice;
 
 import static org.assertj.core.api.Assertions.assertThat;
