+++ src/test/java/racing/RacingCarTest.java
@@ -2,6 +2,7 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.RacingCar;
 
 import static org.assertj.core.api.Assertions.assertThat;
