+++ src/test/java/racing/RacingGameTest.java
@@ -2,6 +2,8 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.controller.RacingGame;
+import racing.domain.RacingGameResult;
 
 import java.util.List;
